requirements.txt
django
psycopg2-binary
pillow
djangorestframework
drf-yasg
python-telegram-bot==13.7